+++ .env.sample
@@ -12,3 +12,6 @@ MODE=development| production
 PGADMIN_EMAIL=admin@admin.com
 PGADMIN_PASSWORD=secret
 PGADMIN_LISTEN_PORT=80
+
+VITE_API_HOST=localhost
+VITE_API_SERVER_PORT=3001
